refactor: split VitePress diagnostic helpers
