+++ crates/vize_maestro/src/ide/diagnostics/collectors.rs
@@ -317,7 +317,6 @@ impl DiagnosticService {
 
         let allocator = vize_carton::Bump::new();
         let (_, errors) = vize_armature::parse(&allocator, &template.content);
-
         errors
             .iter()
             .filter(|error| !error.is_compatibility_notice())
